[refactor] DocumentHeader.js => organisms으로 변경
1. DocumentHeader.js는 내부에 molecules인 DocumentLinkButton로 리스트를 만들어 보여주기 때문에 molecules보다 organisms에 어울린다고 생각해서 변경
